clean and rename
